Fix issue with token-based locking
Previously while you were blocked on getting the token, you could not
make progress even if something arrived into your queue. This changes
it so that we block on either getting a lock or getting a new message
from the queue, whichever comes first.

Also instead of using a queue of tokens, this changes it to use Lock
objects provided by asyncio.