Fix iscoroutine check for pubsub proxied methods (#297)
* Fix iscoroutine check for pubsub proxied methods

Adjust the proxying approach in `RedisPubSubChannelLayer` to ensure the
proxied methods are properly marked as coroutines according to
`inspect.iscoroutinefunction`. This can be issue during testing when
trying to mock channel layer methods.

The underlying issue is a [Python bug][1] where partial methods are not
fully unwrapped to determine if they are coroutines. The workaround
simply moves the proxy wrapper to a top-level async function, tricking
`inspect.iscoroutinefunction` to correctly detecting the proxied async
methods.

[1]: https://bugs.python.org/issue38364

Co-authored-by: Carlton Gibson <carlton.gibson@noumenal.es>